change module name and moving extraction method
